Add tests for basic run
