+++ internal/version/version.go
@@ -1,3 +1,8 @@
 package version
 
+const (
+	ProjectName = "Wallarm API-Firewall"
+	Namespace   = "apifw"
+)
+
 var Version = "develop"
